+++ src/app/dashboard/admin/page.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 
 function page() {
   return (
-    <div className="ml-64 p-6" style={{position: 'relative', }}> {/* Add margin-left and padding */}
+    <div className="ml-64 p-6"> {/* Add margin-left and padding */}
       <h1 className="text-3xl font-bold"> {/* Add styling for the heading */}
-        This is my dashboard and it is what am working on for doing a lot of thin
+        This is my dashboard and it is what am working on for doing a lot of thing am in a page right now
       </h1>
     </div>
   );
